requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.1
aiohttp-retry==2.9.1
aiosignal==1.4.0
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
fastapi==0.120.1
frozenlist==1.8.0
h11==0.16.0
idna==3.11
multidict==6.7.0
propcache==0.4.1
pydantic==2.12.3
pydantic-core==2.41.4
pyjwt==2.10.1
python-dotenv==1.2.1
requests==2.32.5
sniffio==1.3.1
starlette==0.48.0
twilio==9.8.4
typing-extensions==4.15.0
typing-inspection==0.4.2
urllib3==2.5.0
uvicorn==0.38.0
yarl==1.22.0
